requirements.txt
requests==2.32.3
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.9
langchain-text-splitters==0.2.1
html2text==2024.2.26
numpy==1.26.4
neo4j==5.21.0
pydantic==2.7.4
sentence-transformers==3.0.1
chromadb==0.5.3
groq==0.9.0
python-dotenv==1.0.1
gradio==4.36.1
